WebApps/Templates/Admin/admin_sys/requirements.txt
asgiref==3.8.1
backports-datetime-fromisoformat==2.0.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
data-wizard==2.0.0
Django==5.2.3
django-anymail==13.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
environs==14.2.0
et_xmlfile==2.0.0
html-json-forms==1.1.1
idna==3.10
itertable==2.2.0
marshmallow==4.0.0
natural-keys==2.1.1
openpyxl==3.1.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
